requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
pillow==10.4.0
sqlparse==0.5.0
typing_extensions==4.12.2
